@@ -2,11 +2,11 @@ name: tentura
 description: 'Social network for communities'
 publish_to: 'none'
 
-version: 1.3.1
+version: 1.3.6
 
 environment:
   sdk: '>=3.9.2 <4.0.0'
-  flutter: ^3.35.4
+  flutter: ^3.35.5
 
 dependencies:
   flutter:
@@ -20,7 +20,7 @@ dependencies:
   uuid: ^4.4.2
   get_it: ^8.2.0
   latlong2: ^0.9.1
-  injectable: ^2.5.1
+  injectable: ^2.5.2
   auto_route: ^10.1.2
   collection: ^1.19.1
   built_value: ^8.11.1
@@ -35,33 +35,33 @@ dependencies:
   web: ^1.1.1
   http: ^1.5.0
   meta: ^1.16.0
-  drift: ^2.28.1
+  drift: ^2.28.2
   mailto: ^2.0.0
   geocoding: ^4.0.0
-  drift_dev: ^2.28.2
+  drift_dev: ^2.28.3
   geolocator: ^14.0.2
   share_plus: ^12.0.0
   image_picker: ^1.2.0
   url_launcher: ^6.3.2
-  drift_flutter: ^0.2.6
-  firebase_core: ^4.1.0
+  drift_flutter: ^0.2.7
+  firebase_core: ^4.1.1
   share_handler: ^0.0.25
-  mobile_scanner: ^7.0.1
-  firebase_messaging: ^16.0.1
+  mobile_scanner: ^7.1.2
+  firebase_messaging: ^16.0.2
   flutter_native_splash: ^2.4.6
   flutter_secure_storage: ^10.0.0-beta.4
 
   # Logs & Stats
-  logger: ^2.6.1
-  sentry_drift: ^9.6.0
-  sentry_flutter: ^9.6.0
+  logger: ^2.6.2
+  sentry_drift: ^9.7.0
+  sentry_flutter: ^9.7.0
 
   # Remote API related
   ferry: ^0.16.1
   gql_exec: ^1.0.0
-  gql_http_link: ^1.1.0
-  gql_error_link: ^1.0.0
-  gql_dedupe_link: ^2.0.3
+  gql_http_link: ^1.2.0
+  gql_error_link: ^1.0.1
+  gql_dedupe_link: ^4.0.0
   ed25519_edwards: ^0.3.1
   dart_jsonwebtoken: ^3.2.0
   web_socket_client: ^0.2.1
@@ -72,10 +72,10 @@ dependencies:
   image: ^4.5.4
   readmore: ^3.0.0
   qr_flutter: ^4.1.0
-  flutter_map: ^8.2.1
+  flutter_map: ^8.2.2
   flutter_svg: ^2.2.1
-  google_fonts: ^6.3.1
-  validatorless: ^1.2.4
+  google_fonts: ^6.3.2
+  validatorless: ^1.2.5
   blurhash_shader: ^1.1.2
   vector_graphics: ^1.1.19
   visibility_detector: ^0.4.0
@@ -91,7 +91,7 @@ dev_dependencies:
   freezed: ^3.2.3
   build_runner: ^2.7.1
   flutter_lints: ^6.0.0
-  ferry_generator: ^0.14.0-dev.1
+  ferry_generator: ^0.14.0-dev.2
   json_serializable: ^6.11.1
   sentry_dart_plugin: ^3.0.0
   very_good_analysis: ^10.0.0
