@@ -6,83 +6,64 @@ import 'package:tentura/ui/utils/ui_utils.dart';
 
 import '../bloc/beacon_create_cubit.dart';
 
-class ImageTab extends StatefulWidget {
+class ImageTab extends StatelessWidget {
   const ImageTab({super.key});
 
   @override
-  State<ImageTab> createState() => _ImageTabState();
-}
+  Widget build(BuildContext context) {
+    final cubit = context.read<BeaconCreateCubit>();
+    return BlocSelector<BeaconCreateCubit, BeaconCreateState, ImageEntity?>(
+      bloc: cubit,
+      selector: (state) => state.image,
+      builder: (context, image) {
+        const imagePadding = kSpacingLarge * 2;
+        return ListView(
+          children: [
+            // Image Control
+            if (image == null)
+              ListTile(
+                title: Text(L10n.of(context)!.attachImage),
+                trailing: const Icon(Icons.add_a_photo_rounded),
+                onTap: cubit.pickImage,
+              )
+            else
+              ListTile(
+                title: Text(image.fileName),
+                trailing: IconButton(
+                  icon: const Icon(Icons.cancel_rounded),
+                  onPressed: cubit.clearImage,
+                ),
+              ),
 
-class _ImageTabState extends State<ImageTab> {
-  final _controller = TextEditingController();
-
-  late final _l10n = L10n.of(context)!;
-
-  late final _cubit = context.read<BeaconCreateCubit>();
-
-  @override
-  void dispose() {
-    _controller.dispose();
-    super.dispose();
-  }
-
-  @override
-  Widget build(BuildContext context) => ListView(
-    children: [
-      // Image Control
-      Padding(
-        padding: kPaddingSmallV,
-        child: TextFormField(
-          readOnly: true,
-          controller: _controller,
-          decoration: InputDecoration(
-            hintText: _l10n.attachImage,
-            suffixIcon:
-                BlocSelector<
-                  BeaconCreateCubit,
-                  BeaconCreateState,
-                  ImageEntity?
-                >(
-                  selector: (state) => state.image,
-                  builder: (_, image) => image == null
-                      ? const Icon(Icons.add_a_photo_rounded)
-                      : IconButton(
-                          icon: const Icon(Icons.cancel_rounded),
-                          onPressed: () {
-                            _controller.clear();
-                            _cubit.clearImage();
-                          },
+            // Image Container
+            Padding(
+              padding: const EdgeInsets.only(
+                top: kSpacingMedium,
+                bottom: imagePadding,
+                left: imagePadding,
+                right: imagePadding,
+              ),
+              child: image?.imageBytes == null
+                  ? DecoratedBox(
+                      decoration: BoxDecoration(
+                        border: Border.all(
+                          color: Colors.black12,
                         ),
-                ),
-          ),
-          onTap: _cubit.pickImage,
-        ),
-      ),
-
-      // Image Container
-      Padding(
-        padding: const EdgeInsets.all(kSpacingLarge * 2),
-        child: BlocSelector<BeaconCreateCubit, BeaconCreateState, ImageEntity?>(
-          selector: (state) => state.image,
-          builder: (_, image) => image?.imageBytes == null
-              ? DecoratedBox(
-                  decoration: BoxDecoration(
-                    border: Border.all(
-                      color: Colors.black12,
+                      ),
+                      child: const Icon(
+                        Icons.photo_outlined,
+                        size: 256,
+                      ),
+                    )
+                  : Image.memory(
+                      image!.imageBytes!,
+                      key: ObjectKey(image),
+                      fit: BoxFit.fitWidth,
                     ),
-                  ),
-                  child: const Icon(
-                    Icons.photo_outlined,
-                    size: 256,
-                  ),
-                )
-              : Image.memory(
-                  image!.imageBytes!,
-                  key: ObjectKey(image),
-                  fit: BoxFit.fitWidth,
-                ),
-        ),
-      ),
-    ],
-  );
+            ),
+          ],
+        );
+      },
+    );
+  }
 }
