@@ -1,12 +1,14 @@
 import 'package:flutter/material.dart';
 
+import 'package:tentura/consts.dart';
 import 'package:tentura/ui/l10n/l10n.dart';
+import 'package:tentura/ui/utils/string_input_validator.dart';
 
 class ContextAddDialog extends StatefulWidget {
   static Future<String?> show(BuildContext context) =>
       showAdaptiveDialog<String>(
         context: context,
-        builder: (context) => const ContextAddDialog(),
+        builder: (_) => const ContextAddDialog(),
       );
 
   const ContextAddDialog({super.key});
@@ -15,7 +17,8 @@ class ContextAddDialog extends StatefulWidget {
   State<ContextAddDialog> createState() => _ContextAddDialogState();
 }
 
-class _ContextAddDialogState extends State<ContextAddDialog> {
+class _ContextAddDialogState extends State<ContextAddDialog>
+    with StringInputValidator {
   final _controller = TextEditingController();
 
   late final _l10n = L10n.of(context)!;
@@ -29,7 +32,13 @@ class _ContextAddDialogState extends State<ContextAddDialog> {
   @override
   Widget build(BuildContext context) => AlertDialog.adaptive(
     title: Text(_l10n.addNewTopic),
-    content: TextField(controller: _controller),
+    content: TextFormField(
+      autofocus: true,
+      controller: _controller,
+      maxLength: kTitleMaxLength,
+      validator: (value) => titleValidator(_l10n, value),
+      onTapOutside: (_) => FocusManager.instance.primaryFocus?.unfocus(),
+    ),
     actions: [
       TextButton(
         onPressed: () => Navigator.of(context).pop(_controller.text.trim()),
