@@ -2,8 +2,6 @@ import 'dart:typed_data';
 
 Future<Uint8List> readBodyAsBytes(Stream<List<int>> stream) async {
   final builder = BytesBuilder(copy: false);
-  await for (final part in stream) {
-    builder.add(part);
-  }
+  await stream.forEach(builder.add);
   return builder.takeBytes();
 }
